added seed file
